[Permissions] Fix versioned code
