+++ android/expoview/src/main/java/host/exp/exponent/experience/MultipleVersionReactNativeActivity.java
@@ -18,10 +18,19 @@ public class MultipleVersionReactNativeActivity extends ReactNativeActivity impl
     // END_SDK_35
     // BEGIN_SDK_36
     abi36_0_0.com.facebook.react.modules.core.DefaultHardwareBackBtnHandler,
+    abi36_0_0.com.facebook.react.modules.core.PermissionAwareActivity,
     // END_SDK_36
     // ADD_NEW_SDKS_HERE
     // WHEN_PREPARING_SHELL_REMOVE_TO_HERE
     // WHEN_DISTRIBUTING_REMOVE_TO_HERE
     com.facebook.react.modules.core.DefaultHardwareBackBtnHandler {
+
+    // PERMISSION_AWARE_ACTIVITY_IMPLEMENTATION_SDK_36
+    @Override
+    public void requestPermissions(String[] strings, int i, abi36_0_0.com.facebook.react.modules.core.PermissionListener permissionListener) {
+      super.requestPermissions(strings, i, permissionListener::onRequestPermissionsResult);
+    }
+    // END_PERMISSION_AWARE_ACTIVITY_IMPLEMENTATION_SDK_36
+    // ADD_NEW_PERMISSION_AWARE_ACTIVITY_IMPLEMENTATION
 
 }
